Refactor dataset and IO modules for clearer separation
- Move dataset loading and writing utilities to lisbet/io
- Move external dataset sources to lisbet/io/ext_sources
- Remove redundant datasets/__init__.py and training/io.py
- Update imports throughout codebase to use new module structure
- Refactor WindowDataset to use dict-style access for posetracks
- Remove obsolete input pipeline tests
