+++ src/lisbet/datasets.py
@@ -105,8 +105,8 @@ class WindowDataset(IterableDataset):
         if window_size is None:
             window_size = self.window_size
 
-        x = self.records[curr_key].posetracks
-        seq_len = x.sizes["time"]
+        x_data = self.records[curr_key]["posetracks"]
+        seq_len = x_data.sizes["time"]
 
         # Compute actual window size
         act_window_size = int(np.rint(self.fps_scaling * window_size))
@@ -128,10 +128,14 @@ class WindowDataset(IterableDataset):
         # logging.debug("Data bounds: (%d, %d, %d)", start_idx, curr_loc, stop_idx)
 
         # Pad data with zeros
-        past_pad = np.zeros((past_n, x.sizes["features"]))
-        future_pad = np.zeros((future_n, x.sizes["features"]))
-        x = np.concatenate(
-            [past_pad, x["position"].isel(time=slice(start_idx, stop_idx)), future_pad],
+        past_pad = np.zeros((past_n, x_data.sizes["features"]))
+        future_pad = np.zeros((future_n, x_data.sizes["features"]))
+        x_data = np.concatenate(
+            [
+                past_pad,
+                x_data["position"].isel(time=slice(start_idx, stop_idx)),
+                future_pad,
+            ],
             axis=0,
         )
 
@@ -140,10 +144,10 @@ class WindowDataset(IterableDataset):
         # ), f"{seq_len}, {start_idx}, {curr_loc}, {stop_idx}, {past_n}, {future_n}"
 
         # Interpolate frames to get exactly window_size frames
-        f1d = interp1d(np.linspace(0, 1, act_window_size), x, axis=0)
-        x = f1d(np.linspace(0, 1, window_size))
+        f1d = interp1d(np.linspace(0, 1, act_window_size), x_data, axis=0)
+        x_data = f1d(np.linspace(0, 1, window_size))
 
-        return x
+        return x_data
 
 
 class RandomWindowDataset(WindowDataset):
