--- src/lisbet/datasets/__init__.py
@@ -1,4 +1,0 @@
-from .core import dump_records, load_records
-
-__all__ = ["dump_records", "load_records"]
-__doc__ = """Datasets input for LISBET."""
