+++ src/lisbet/io/__init__.py
@@ -1,0 +1,18 @@
+from .core import (
+    dump_model_config,
+    dump_profiling_results,
+    dump_records,
+    dump_weights,
+    load_multi_records,
+    load_records,
+)
+
+__all__ = [
+    "dump_model_config",
+    "dump_profiling_results",
+    "dump_records",
+    "dump_weights",
+    "load_multi_records",
+    "load_records",
+]
+__doc__ = """Datasets input for LISBET."""
